refactor: move starknet specific code to directory
