+++ apps/api/src/starknet/abis/l1/L1AvatarExectionStrategy.json
