+++ apps/api/src/starknet/abis/l1/SimpleQuorumExecutionStrategy.json
